editoast: move InfraError to generated_data
