+++ editoast/src/generated_data/error/buffer_stops.rs
@@ -1,10 +1,10 @@
 use super::GlobalErrorGenerator;
 use super::NoContext;
 use crate::generated_data::error::ObjectErrorGenerator;
+use crate::generated_data::infra_error::InfraError;
 use crate::infra_cache::Graph;
 use crate::infra_cache::InfraCache;
 use crate::infra_cache::ObjectCache;
-use crate::schema::InfraError;
 use editoast_schemas::infra::Endpoint;
 use editoast_schemas::infra::TrackEndpoint;
 use editoast_schemas::primitives::ObjectRef;
