+++ editoast/src/generated_data/error/routes.rs
@@ -3,10 +3,10 @@ use std::collections::HashSet;
 
 use super::GlobalErrorGenerator;
 use crate::generated_data::error::ObjectErrorGenerator;
+use crate::generated_data::infra_error::InfraError;
 use crate::infra_cache::Graph;
 use crate::infra_cache::InfraCache;
 use crate::infra_cache::ObjectCache;
-use crate::schema::InfraError;
 use editoast_schemas::infra::Waypoint;
 use editoast_schemas::primitives::OSRDIdentified;
 use editoast_schemas::primitives::OSRDObject;
