+++ editoast/src/generated_data/error/speed_sections.rs
@@ -6,10 +6,10 @@ use rangemap::RangeMap;
 use super::GlobalErrorGenerator;
 use super::NoContext;
 use crate::generated_data::error::ObjectErrorGenerator;
+use crate::generated_data::infra_error::InfraError;
 use crate::infra_cache::Graph;
 use crate::infra_cache::InfraCache;
 use crate::infra_cache::ObjectCache;
-use crate::schema::InfraError;
 use editoast_schemas::infra::ApplicableDirections;
 use editoast_schemas::infra::Direction;
 use editoast_schemas::primitives::OSRDIdentified;
