+++ editoast/src/schema/mod.rs
@@ -1,10 +1,7 @@
-mod errors;
 mod railjson;
 pub mod track_section;
 pub mod v2;
 
-pub use errors::InfraError;
-pub use errors::InfraErrorType;
 pub use railjson::RailJson;
 pub use railjson::RAILJSON_VERSION;
 pub use track_section::TrackSection;
