+++ editoast/src/views/infra/auto_fixes/mod.rs
@@ -15,6 +15,7 @@ use tracing::error;
 use crate::error::InternalError;
 use crate::error::Result;
 use crate::generated_data::generate_infra_errors;
+use crate::generated_data::infra_error::InfraError;
 use crate::infra_cache::operation::CacheOperation;
 use crate::infra_cache::operation::DeleteOperation;
 use crate::infra_cache::operation::Operation;
@@ -24,7 +25,6 @@ use crate::infra_cache::InfraCache;
 use crate::infra_cache::ObjectCache;
 use crate::modelsv2::prelude::*;
 use crate::modelsv2::Infra;
-use crate::schema::InfraError;
 use crate::views::infra::InfraApiError;
 use crate::views::infra::InfraIdParam;
 use crate::DbPool;
