+++ editoast/src/views/infra/auto_fixes/route.rs
@@ -4,8 +4,8 @@ use tracing::debug;
 
 use super::new_ref_fix_delete_pair;
 use super::Fix;
-use crate::schema::InfraError;
-use crate::schema::InfraErrorType;
+use crate::generated_data::infra_error::InfraError;
+use crate::generated_data::infra_error::InfraErrorType;
 use editoast_schemas::infra::Route;
 use editoast_schemas::primitives::OSRDIdentified as _;
 use editoast_schemas::primitives::OSRDObject as _;
